Speed Manager scripts and Closest Human Pointer
Today, I have added two different speed manager script, one that calculates a central one for zombies, and a second that calculates for humans. Then, I have added arrow to the UI, that shows the closest central human to the player making it easier for player to find the zombies in the map.